more cmake, reorganization
