+++ server/enclave/CMakeLists.txt
@@ -1,7 +1,11 @@
 # Use the edger8r to generate C bindings from the EDL file.
-add_custom_command(OUTPUT modelaggregator_t.h modelaggregator_t.c modelaggregator_args.h
+add_custom_command(
+  OUTPUT modelaggregator_t.h modelaggregator_t.c modelaggregator_args.h
   DEPENDS ${CMAKE_SOURCE_DIR}/modelaggregator.edl
-  COMMAND openenclave::oeedger8r --trusted ${CMAKE_SOURCE_DIR}/modelaggregator.edl)
+  COMMAND
+    openenclave::oeedger8r --trusted ${CMAKE_SOURCE_DIR}/modelaggregator.edl
+    --search-path ${OE_INCLUDEDIR} --search-path
+    ${OE_INCLUDEDIR}/openenclave/edl/sgx)
 
 add_executable(enclave ecalls.cpp ${CMAKE_CURRENT_BINARY_DIR}/modelaggregator_t.c)
