+++ server/host/CMakeLists.txt
@@ -1,6 +1,11 @@
-add_custom_command(OUTPUT modelaggregator_u.h modelaggregator_u.c modelaggregator_args.h
+# Use the edger8r to generate C bindings from the EDL file.
+add_custom_command(
+  OUTPUT modelaggregator_u.h modelaggregator_u.c modelaggregator_args.h
   DEPENDS ${CMAKE_SOURCE_DIR}/modelaggregator.edl
-  COMMAND openenclave::oeedger8r --untrusted ${CMAKE_SOURCE_DIR}/modelaggregator.edl)
+  COMMAND
+    openenclave::oeedger8r --untrusted ${CMAKE_SOURCE_DIR}/modelaggregator.edl
+    --search-path ${OE_INCLUDEDIR} --search-path
+    ${OE_INCLUDEDIR}/openenclave/edl/sgx)
 
 add_library(modelaggregator_host host.cpp ${CMAKE_CURRENT_BINARY_DIR}/modelaggregator_u.c)
