+++ server/host/host.cpp
@@ -1,15 +1,15 @@
 #include <openenclave/host.h>
 #include <stdio.h>
 
-#include "../enclave.h"
+#include "enclave.h"
 
 // Include the untrusted modelaggregator header that is generated
 // during the build. This file is generated by calling the
 // sdk tool oeedger8r against the modelaggregator.edl file.
 #include "modelaggregator_u.h"
 
-char* path;
-uint32_t flags;
+char* path = "../enclave/helloworldenc.signed";
+uint32_t flags = OE_ENCLAVE_FLAG_DEBUG | OE_ENCLAVE_FLAG_SIMULATE;
 
 // This is the function that the Python code will call into
 // Returns NULL on failure, new encrypted model on success
