+++ server/tests/CMakeLists.txt
@@ -2,6 +2,6 @@ include_directories (../)
 
 set(TEST_SRCS host/ enclave/)
 
-add_executable(Test tests.cpp)
+add_executable(Test host_test.cpp)
 
 target_link_libraries(Test modelaggregator_host)
