+++ server/tests/host_test.cpp
@@ -1,5 +1,7 @@
 #include <stdio.h>
 
+#include "../host/host.cpp"
+
 int main(int argc, char* argv[]) 
 {
     printf("builds!");
